Latitude CF axis object

This class represents a latitude axis. Its values are numeric. This class adds some logic that is specific to latitudes, such as their range, orientation and meaning.

Super classes

CFObject -> CFData -> CFAxis -> CFAxisNumeric -> CFAxisLatitude

Active bindings

friendlyClassName

(read-only) A nice description of the class.

Methods

Public methods

CFAxisLatitude$new()

CFAxisLatitude$copy()

CFAxisLatitude$copy_with_values()

CFAxisLatitude$subset()

CFAxisLatitude$append()

Inherited methods

CFObject$append_attribute()

CFObject$attribute()

CFObject$attributes_identical()

CFObject$delete_attribute()

CFObject$print_attributes()

CFObject$set_attribute()

CFObject$write_attributes()

CFData$dim()

CFData$read_chunk()

CFData$read_data()

CFAxis$attach_to_group()

CFAxis$can_append()

CFAxis$configure_terms()

CFAxis$copy_terms()

CFAxis$detach()

CFAxis$geozarr_axis()

CFAxis$geozarr_coordinates()

CFAxis$peek()

CFAxis$shard()

CFAxis$write()

CFAxisNumeric$brief()

CFAxisNumeric$identical()

CFAxisNumeric$indexOf()

CFAxisNumeric$print()

CFAxisNumeric$range()

CFAxisNumeric$slice()

CFAxisLatitude$new()

Create a new instance of this class.

Creating a new latitude axis is more easily done with the makeLatitudeAxis() function.

Usage

CFAxisLatitude$new(
  var,
  group,
  values,
  start = 1L,
  count = NA,
  attributes = data.frame()
)

Arguments

var

The name of the axis when creating a new axis. When reading an axis from file, the NCVariable object that describes this instance.

group

The CFGroup that this instance will live in.

values

Optional. The values of the axis in a vector. The values have to be numeric within the range (-90, 90) and monotonic. Ignored when argument var is a NCVariable object.

start

Optional. Integer index where to start reading axis data from file. The index may be NA to start reading data from the start.

count

Optional. Number of elements to read from file. This may be NA to read to the end of the data.

attributes

Optional. A data.frame with the attributes of the axis. When an empty data.frame (default) and argument var is an NCVariable instance, attributes of the axis will be taken from the netCDF resource.

CFAxisLatitude$copy()

Create a copy of this axis. The copy is completely separate from self, meaning that both self and all of its components are made from new instances.

Usage

CFAxisLatitude$copy(name = "", group)

Arguments

name

The name for the new axis. If an empty string is passed, will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

Returns

The newly created axis.

CFAxisLatitude$copy_with_values()

Create a copy of this axis but using the supplied values. The attributes are copied to the new axis. Boundary values and auxiliary coordinates are not copied.

After this operation the attributes of the newly created axes may not be accurate, except for the "actual_range" attribute. The calling code should set, modify or delete attributes as appropriate.

Usage

CFAxisLatitude$copy_with_values(name = "", group, values)

Arguments

name

The name for the new axis. If an empty string is passed, will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

values

The values to the used with the copy of this axis.

Returns

The newly created axis.

CFAxisLatitude$subset()

Return a latitude axis spanning a smaller coordinate range. This method returns an axis which spans the range of indices given by the rng argument.

Usage

CFAxisLatitude$subset(name = "", group, rng = NULL)

Arguments

name

The name for the new axis. If an empty string is passed (default), will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

rng

The range of indices whose values from this axis to include in the returned axis. If the value of the argument is NULL, return a copy of the axis.

Returns

A new CFAxisLatitude instance covering the indicated range of indices. If the value of the argument rng is NULL, return a copy of self as the new axis.

CFAxisLatitude$append()

Append a vector of values at the end of the current values of the axis. Boundary values are appended as well but if either this axis or the from axis does not have boundary values, neither will the resulting axis.

Usage

CFAxisLatitude$append(from)

Arguments

from

An instance of CFAxisLatitude whose values to append to the values of this axis.

group

The CFGroup where the copy of this axis will live.

Returns

A new CFAxisLatitude instance with values from this axis and the from axis appended.