Longitude CF axis object

This class represents a longitude axis. Its values are numeric. This class is used for axes that represent longitudes. This class adds some logic that is specific to longitudes, such as their range, orientation and their meaning. (In the near future, it will also support selecting data that crosses the 0-360 degree boundary.)

Super classes

ncdfCF::CFObject -> ncdfCF::CFData -> ncdfCF::CFAxis -> ncdfCF::CFAxisNumeric -> CFAxisLongitude

Active bindings

friendlyClassName

(read-only) A nice description of the class.

Methods

Public methods

CFAxisLongitude$new()

CFAxisLongitude$copy()

CFAxisLongitude$copy_with_values()

CFAxisLongitude$subset()

CFAxisLongitude$append()

Inherited methods

ncdfCF::CFObject$append_attribute()

ncdfCF::CFObject$attribute()

ncdfCF::CFObject$attributes_identical()

ncdfCF::CFObject$delete_attribute()

ncdfCF::CFObject$print_attributes()

ncdfCF::CFObject$set_attribute()

ncdfCF::CFObject$write_attributes()

ncdfCF::CFData$dim()

ncdfCF::CFAxis$attach_to_group()

ncdfCF::CFAxis$can_append()

ncdfCF::CFAxis$configure_terms()

ncdfCF::CFAxis$copy_terms()

ncdfCF::CFAxis$detach()

ncdfCF::CFAxis$peek()

ncdfCF::CFAxis$shard()

ncdfCF::CFAxis$write()

ncdfCF::CFAxisNumeric$brief()

ncdfCF::CFAxisNumeric$identical()

ncdfCF::CFAxisNumeric$indexOf()

ncdfCF::CFAxisNumeric$print()

ncdfCF::CFAxisNumeric$range()

ncdfCF::CFAxisNumeric$slice()

Method new()

Create a new instance of this class.

Creating a new longitude axis is more easily done with the makeLongitudeAxis() function.

Usage

CFAxisLongitude$new(
  var,
  group,
  values,
  start = 1L,
  count = NA,
  attributes = data.frame()
)

Arguments

var

The name of the axis when creating a new axis. When reading an axis from file, the NCVariable object that describes this instance.

group

The CFGroup that this instance will live in.

values

Optional. The values of the axis in a vector. The values have to be numeric with the maximum value no larger than the minimum value + 360, and monotonic. Ignored when argument var is a NCVariable object.

start

Optional. Integer index where to start reading axis data from file. The index may be NA to start reading data from the start.

count

Optional. Number of elements to read from file. This may be NA to read to the end of the data.

attributes

Optional. A data.frame with the attributes of the axis. When an empty data.frame (default) and argument var is an NCVariable instance, attributes of the axis will be taken from the netCDF resource.

Method copy()

Create a copy of this axis. The copy is completely separate from self, meaning that both self and all of its components are made from new instances.

Usage

CFAxisLongitude$copy(name = "", group)

Arguments

name

The name for the new axis. If an empty string is passed, will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

Returns

The newly created axis.

Method copy_with_values()

Create a copy of this axis but using the supplied values. The attributes are copied to the new axis. Boundary values and auxiliary coordinates are not copied.

After this operation the attributes of the newly created axes may not be accurate, except for the "actual_range" attribute. The calling code should set, modify or delete attributes as appropriate.

Usage

CFAxisLongitude$copy_with_values(name = "", group, values)

Arguments

name

The name for the new axis. If an empty string is passed, will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

values

The values to the used with the copy of this axis.

Returns

The newly created axis.

Method subset()

Return a longitude axis spanning a smaller coordinate range. This method returns an axis which spans the range of indices given by the rng argument.

Usage

CFAxisLongitude$subset(name = "", group, rng = NULL)

Arguments

name

The name for the new axis. If an empty string is passed (default), will use the name of this axis.

group

The CFGroup where the copy of this axis will live.

rng

The range of indices whose values from this axis to include in the returned axis. If the value of the argument is NULL, return a copy of the axis.

Returns

A new CFAxisLongitude instance covering the indicated range of indices. If the value of the argument rng is NULL, return a copy of self as the new axis.

Method append()

Append a vector of values at the end of the current values of the axis. Boundary values are appended as well but if either this axis or the from axis does not have boundary values, neither will the resulting axis.

Usage

CFAxisLongitude$append(from)

Arguments

from

An instance of CFAxisLongitude whose values to append to the values of this axis.

group

The CFGroup where the copy of this axis will live.

Returns

A new CFAxisLongitude instance with values from this axis and the from axis appended.