NetCDF user-defined type

This class represents user-defined types in a netCDF file. Interpretation of the UDT typically requires knowledge of the data set or application.

Super class

ncdfCF::NCObject -> NCUDT

Public fields

clss

The class of the UDT, one of "builtin", "compound", "enum", "opaque", or "vlen".

size

Size in bytes of a single item of the type (or a single element of a "vlen").

basetype

Name of the netCDF base type of each element ("enum" and "vlen" only).

value

Named vector with numeric values of all members ("enum" only).

offset

Named vector with the offset of each field in bytes from the beginning of the "compound" type.

subtype

Named vector with the netCDF base type name of each field of a "compound" type.

dimsizes

Named list with array dimensions of each field of a "compound" type. A NULL length indicates a scalar.

Methods

Public methods

NCUDT$new()

NCUDT$clone()

Inherited methods

ncdfCF::NCObject$add_coordinates()

ncdfCF::NCObject$append_attribute()

ncdfCF::NCObject$attribute()

ncdfCF::NCObject$delete_attribute()

ncdfCF::NCObject$print_attributes()

ncdfCF::NCObject$set_attribute()

ncdfCF::NCObject$write_attributes()

Method new()

Create a new netCDF user-defined type. This class represents a user-defined type. It is instantiated when opening a netCDF resource.

Usage

NCUDT$new(id, name, clss, size, basetype, value, offset, subtype, dimsizes)

Arguments

id

Numeric identifier of the user-defined type.

name

Character string with the name of the user-defined type.

clss

The class of the UDT, one of "builtin", "compound", "enum", "opaque", or "vlen".

size

Size in bytes of a single item of the type (or a single element of a "vlen").

basetype

Name of the netCDF base type of each element ("enum" and "vlen" only).

value

Named vector with numeric values of all members ("enum" only).

offset

Named vector with the offset of each field in bytes from the beginning of the "compound" type.

subtype

Named vector with the netCDF base type name of each field of a "compound" type.

dimsizes

Named list with array dimensions of each field of a "compound" type. A NULL length indicates a scalar.

Returns

An instance of this class.

Method clone()

The objects of this class are cloneable with this method.

Usage

NCUDT$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.