Names or axis values of an CF object

Retrieve the variable or axis names of an ncdfCF object. The names() function gives the names of the variables in the data set, preceded by the path to the group if the resource uses groups. The return value of the dimnames() function differs depending on the type of object:

CFDataset, CFVariable: The dimnames are returned as a vector of the names of the axes of the data set or variable, preceded with the path to the group if the resource uses groups. Note that this differs markedly from the base::dimnames() functionality.

CFAxisNumeric, CFAxisLongitude, CFAxisLatitude, CFAxisVertical: The coordinate values along the axis as a numeric vector.

CFAxisTime: The coordinate values along the axis as a character vector containing timestamps in ISO8601 format. This could be dates or date-times if time information is available in the axis.

CFAxisCharacter: The coordinate values along the axis as a character vector.

CFAxisDiscrete: The index values of the axis, either along the entire axis, or a portion thereof.

Usage

# S3 method for class 'CFDataset'
names(x)

Arguments

x

An CFObject whose axis names to retrieve. This could be CFDataset, CFVariable, or a class descending from CFAxis.

Value

A vector as described in the Description section.

Examples

fn <- system.file("extdata",
  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20230101-20231231_vncdfCF.nc",
  package = "ncdfCF")
ds <- open_ncdf(fn)

# Names of data variables
names(ds)
#> [1] "pr"

# CFDataset
dimnames(ds)
#> [1] "time" "lon"  "lat" 

# CFVariable
pr <- ds[["pr"]]
dimnames(pr)
#> [1] "lon"  "lat"  "time"

# CFAxisNumeric
lon <- ds[["lon"]]
dimnames(lon)
#>  [1]  5.625000  6.328125  7.031250  7.734375  8.437500  9.140625  9.843750
#>  [8] 10.546875 11.250000 11.953125 12.656250 13.359375 14.062500 14.765625

# CFAxisTime
t <- ds[["time"]]
dimnames(t)
#>   [1] "2023-01-01T12:00:00" "2023-01-02T12:00:00" "2023-01-03T12:00:00"
#>   [4] "2023-01-04T12:00:00" "2023-01-05T12:00:00" "2023-01-06T12:00:00"
#>   [7] "2023-01-07T12:00:00" "2023-01-08T12:00:00" "2023-01-09T12:00:00"
#>  [10] "2023-01-10T12:00:00" "2023-01-11T12:00:00" "2023-01-12T12:00:00"
#>  [13] "2023-01-13T12:00:00" "2023-01-14T12:00:00" "2023-01-15T12:00:00"
#>  [16] "2023-01-16T12:00:00" "2023-01-17T12:00:00" "2023-01-18T12:00:00"
#>  [19] "2023-01-19T12:00:00" "2023-01-20T12:00:00" "2023-01-21T12:00:00"
#>  [22] "2023-01-22T12:00:00" "2023-01-23T12:00:00" "2023-01-24T12:00:00"
#>  [25] "2023-01-25T12:00:00" "2023-01-26T12:00:00" "2023-01-27T12:00:00"
#>  [28] "2023-01-28T12:00:00" "2023-01-29T12:00:00" "2023-01-30T12:00:00"
#>  [31] "2023-01-31T12:00:00" "2023-02-01T12:00:00" "2023-02-02T12:00:00"
#>  [34] "2023-02-03T12:00:00" "2023-02-04T12:00:00" "2023-02-05T12:00:00"
#>  [37] "2023-02-06T12:00:00" "2023-02-07T12:00:00" "2023-02-08T12:00:00"
#>  [40] "2023-02-09T12:00:00" "2023-02-10T12:00:00" "2023-02-11T12:00:00"
#>  [43] "2023-02-12T12:00:00" "2023-02-13T12:00:00" "2023-02-14T12:00:00"
#>  [46] "2023-02-15T12:00:00" "2023-02-16T12:00:00" "2023-02-17T12:00:00"
#>  [49] "2023-02-18T12:00:00" "2023-02-19T12:00:00" "2023-02-20T12:00:00"
#>  [52] "2023-02-21T12:00:00" "2023-02-22T12:00:00" "2023-02-23T12:00:00"
#>  [55] "2023-02-24T12:00:00" "2023-02-25T12:00:00" "2023-02-26T12:00:00"
#>  [58] "2023-02-27T12:00:00" "2023-02-28T12:00:00" "2023-03-01T12:00:00"
#>  [61] "2023-03-02T12:00:00" "2023-03-03T12:00:00" "2023-03-04T12:00:00"
#>  [64] "2023-03-05T12:00:00" "2023-03-06T12:00:00" "2023-03-07T12:00:00"
#>  [67] "2023-03-08T12:00:00" "2023-03-09T12:00:00" "2023-03-10T12:00:00"
#>  [70] "2023-03-11T12:00:00" "2023-03-12T12:00:00" "2023-03-13T12:00:00"
#>  [73] "2023-03-14T12:00:00" "2023-03-15T12:00:00" "2023-03-16T12:00:00"
#>  [76] "2023-03-17T12:00:00" "2023-03-18T12:00:00" "2023-03-19T12:00:00"
#>  [79] "2023-03-20T12:00:00" "2023-03-21T12:00:00" "2023-03-22T12:00:00"
#>  [82] "2023-03-23T12:00:00" "2023-03-24T12:00:00" "2023-03-25T12:00:00"
#>  [85] "2023-03-26T12:00:00" "2023-03-27T12:00:00" "2023-03-28T12:00:00"
#>  [88] "2023-03-29T12:00:00" "2023-03-30T12:00:00" "2023-03-31T12:00:00"
#>  [91] "2023-04-01T12:00:00" "2023-04-02T12:00:00" "2023-04-03T12:00:00"
#>  [94] "2023-04-04T12:00:00" "2023-04-05T12:00:00" "2023-04-06T12:00:00"
#>  [97] "2023-04-07T12:00:00" "2023-04-08T12:00:00" "2023-04-09T12:00:00"
#> [100] "2023-04-10T12:00:00" "2023-04-11T12:00:00" "2023-04-12T12:00:00"
#> [103] "2023-04-13T12:00:00" "2023-04-14T12:00:00" "2023-04-15T12:00:00"
#> [106] "2023-04-16T12:00:00" "2023-04-17T12:00:00" "2023-04-18T12:00:00"
#> [109] "2023-04-19T12:00:00" "2023-04-20T12:00:00" "2023-04-21T12:00:00"
#> [112] "2023-04-22T12:00:00" "2023-04-23T12:00:00" "2023-04-24T12:00:00"
#> [115] "2023-04-25T12:00:00" "2023-04-26T12:00:00" "2023-04-27T12:00:00"
#> [118] "2023-04-28T12:00:00" "2023-04-29T12:00:00" "2023-04-30T12:00:00"
#> [121] "2023-05-01T12:00:00" "2023-05-02T12:00:00" "2023-05-03T12:00:00"
#> [124] "2023-05-04T12:00:00" "2023-05-05T12:00:00" "2023-05-06T12:00:00"
#> [127] "2023-05-07T12:00:00" "2023-05-08T12:00:00" "2023-05-09T12:00:00"
#> [130] "2023-05-10T12:00:00" "2023-05-11T12:00:00" "2023-05-12T12:00:00"
#> [133] "2023-05-13T12:00:00" "2023-05-14T12:00:00" "2023-05-15T12:00:00"
#> [136] "2023-05-16T12:00:00" "2023-05-17T12:00:00" "2023-05-18T12:00:00"
#> [139] "2023-05-19T12:00:00" "2023-05-20T12:00:00" "2023-05-21T12:00:00"
#> [142] "2023-05-22T12:00:00" "2023-05-23T12:00:00" "2023-05-24T12:00:00"
#> [145] "2023-05-25T12:00:00" "2023-05-26T12:00:00" "2023-05-27T12:00:00"
#> [148] "2023-05-28T12:00:00" "2023-05-29T12:00:00" "2023-05-30T12:00:00"
#> [151] "2023-05-31T12:00:00" "2023-06-01T12:00:00" "2023-06-02T12:00:00"
#> [154] "2023-06-03T12:00:00" "2023-06-04T12:00:00" "2023-06-05T12:00:00"
#> [157] "2023-06-06T12:00:00" "2023-06-07T12:00:00" "2023-06-08T12:00:00"
#> [160] "2023-06-09T12:00:00" "2023-06-10T12:00:00" "2023-06-11T12:00:00"
#> [163] "2023-06-12T12:00:00" "2023-06-13T12:00:00" "2023-06-14T12:00:00"
#> [166] "2023-06-15T12:00:00" "2023-06-16T12:00:00" "2023-06-17T12:00:00"
#> [169] "2023-06-18T12:00:00" "2023-06-19T12:00:00" "2023-06-20T12:00:00"
#> [172] "2023-06-21T12:00:00" "2023-06-22T12:00:00" "2023-06-23T12:00:00"
#> [175] "2023-06-24T12:00:00" "2023-06-25T12:00:00" "2023-06-26T12:00:00"
#> [178] "2023-06-27T12:00:00" "2023-06-28T12:00:00" "2023-06-29T12:00:00"
#> [181] "2023-06-30T12:00:00" "2023-07-01T12:00:00" "2023-07-02T12:00:00"
#> [184] "2023-07-03T12:00:00" "2023-07-04T12:00:00" "2023-07-05T12:00:00"
#> [187] "2023-07-06T12:00:00" "2023-07-07T12:00:00" "2023-07-08T12:00:00"
#> [190] "2023-07-09T12:00:00" "2023-07-10T12:00:00" "2023-07-11T12:00:00"
#> [193] "2023-07-12T12:00:00" "2023-07-13T12:00:00" "2023-07-14T12:00:00"
#> [196] "2023-07-15T12:00:00" "2023-07-16T12:00:00" "2023-07-17T12:00:00"
#> [199] "2023-07-18T12:00:00" "2023-07-19T12:00:00" "2023-07-20T12:00:00"
#> [202] "2023-07-21T12:00:00" "2023-07-22T12:00:00" "2023-07-23T12:00:00"
#> [205] "2023-07-24T12:00:00" "2023-07-25T12:00:00" "2023-07-26T12:00:00"
#> [208] "2023-07-27T12:00:00" "2023-07-28T12:00:00" "2023-07-29T12:00:00"
#> [211] "2023-07-30T12:00:00" "2023-07-31T12:00:00" "2023-08-01T12:00:00"
#> [214] "2023-08-02T12:00:00" "2023-08-03T12:00:00" "2023-08-04T12:00:00"
#> [217] "2023-08-05T12:00:00" "2023-08-06T12:00:00" "2023-08-07T12:00:00"
#> [220] "2023-08-08T12:00:00" "2023-08-09T12:00:00" "2023-08-10T12:00:00"
#> [223] "2023-08-11T12:00:00" "2023-08-12T12:00:00" "2023-08-13T12:00:00"
#> [226] "2023-08-14T12:00:00" "2023-08-15T12:00:00" "2023-08-16T12:00:00"
#> [229] "2023-08-17T12:00:00" "2023-08-18T12:00:00" "2023-08-19T12:00:00"
#> [232] "2023-08-20T12:00:00" "2023-08-21T12:00:00" "2023-08-22T12:00:00"
#> [235] "2023-08-23T12:00:00" "2023-08-24T12:00:00" "2023-08-25T12:00:00"
#> [238] "2023-08-26T12:00:00" "2023-08-27T12:00:00" "2023-08-28T12:00:00"
#> [241] "2023-08-29T12:00:00" "2023-08-30T12:00:00" "2023-08-31T12:00:00"
#> [244] "2023-09-01T12:00:00" "2023-09-02T12:00:00" "2023-09-03T12:00:00"
#> [247] "2023-09-04T12:00:00" "2023-09-05T12:00:00" "2023-09-06T12:00:00"
#> [250] "2023-09-07T12:00:00" "2023-09-08T12:00:00" "2023-09-09T12:00:00"
#> [253] "2023-09-10T12:00:00" "2023-09-11T12:00:00" "2023-09-12T12:00:00"
#> [256] "2023-09-13T12:00:00" "2023-09-14T12:00:00" "2023-09-15T12:00:00"
#> [259] "2023-09-16T12:00:00" "2023-09-17T12:00:00" "2023-09-18T12:00:00"
#> [262] "2023-09-19T12:00:00" "2023-09-20T12:00:00" "2023-09-21T12:00:00"
#> [265] "2023-09-22T12:00:00" "2023-09-23T12:00:00" "2023-09-24T12:00:00"
#> [268] "2023-09-25T12:00:00" "2023-09-26T12:00:00" "2023-09-27T12:00:00"
#> [271] "2023-09-28T12:00:00" "2023-09-29T12:00:00" "2023-09-30T12:00:00"
#> [274] "2023-10-01T12:00:00" "2023-10-02T12:00:00" "2023-10-03T12:00:00"
#> [277] "2023-10-04T12:00:00" "2023-10-05T12:00:00" "2023-10-06T12:00:00"
#> [280] "2023-10-07T12:00:00" "2023-10-08T12:00:00" "2023-10-09T12:00:00"
#> [283] "2023-10-10T12:00:00" "2023-10-11T12:00:00" "2023-10-12T12:00:00"
#> [286] "2023-10-13T12:00:00" "2023-10-14T12:00:00" "2023-10-15T12:00:00"
#> [289] "2023-10-16T12:00:00" "2023-10-17T12:00:00" "2023-10-18T12:00:00"
#> [292] "2023-10-19T12:00:00" "2023-10-20T12:00:00" "2023-10-21T12:00:00"
#> [295] "2023-10-22T12:00:00" "2023-10-23T12:00:00" "2023-10-24T12:00:00"
#> [298] "2023-10-25T12:00:00" "2023-10-26T12:00:00" "2023-10-27T12:00:00"
#> [301] "2023-10-28T12:00:00" "2023-10-29T12:00:00" "2023-10-30T12:00:00"
#> [304] "2023-10-31T12:00:00" "2023-11-01T12:00:00" "2023-11-02T12:00:00"
#> [307] "2023-11-03T12:00:00" "2023-11-04T12:00:00" "2023-11-05T12:00:00"
#> [310] "2023-11-06T12:00:00" "2023-11-07T12:00:00" "2023-11-08T12:00:00"
#> [313] "2023-11-09T12:00:00" "2023-11-10T12:00:00" "2023-11-11T12:00:00"
#> [316] "2023-11-12T12:00:00" "2023-11-13T12:00:00" "2023-11-14T12:00:00"
#> [319] "2023-11-15T12:00:00" "2023-11-16T12:00:00" "2023-11-17T12:00:00"
#> [322] "2023-11-18T12:00:00" "2023-11-19T12:00:00" "2023-11-20T12:00:00"
#> [325] "2023-11-21T12:00:00" "2023-11-22T12:00:00" "2023-11-23T12:00:00"
#> [328] "2023-11-24T12:00:00" "2023-11-25T12:00:00" "2023-11-26T12:00:00"
#> [331] "2023-11-27T12:00:00" "2023-11-28T12:00:00" "2023-11-29T12:00:00"
#> [334] "2023-11-30T12:00:00" "2023-12-01T12:00:00" "2023-12-02T12:00:00"
#> [337] "2023-12-03T12:00:00" "2023-12-04T12:00:00" "2023-12-05T12:00:00"
#> [340] "2023-12-06T12:00:00" "2023-12-07T12:00:00" "2023-12-08T12:00:00"
#> [343] "2023-12-09T12:00:00" "2023-12-10T12:00:00" "2023-12-11T12:00:00"
#> [346] "2023-12-12T12:00:00" "2023-12-13T12:00:00" "2023-12-14T12:00:00"
#> [349] "2023-12-15T12:00:00" "2023-12-16T12:00:00" "2023-12-17T12:00:00"
#> [352] "2023-12-18T12:00:00" "2023-12-19T12:00:00" "2023-12-20T12:00:00"
#> [355] "2023-12-21T12:00:00" "2023-12-22T12:00:00" "2023-12-23T12:00:00"
#> [358] "2023-12-24T12:00:00" "2023-12-25T12:00:00" "2023-12-26T12:00:00"
#> [361] "2023-12-27T12:00:00" "2023-12-28T12:00:00" "2023-12-29T12:00:00"
#> [364] "2023-12-30T12:00:00" "2023-12-31T12:00:00"